requirements.txt
discord.py
pomice
pillow